feat: implement a "rubocop version manager"
This is a script that auto-installs and auto-starts different version of
rubocop based on a .rubocop-version file in the packages. This is meant as a
way to smoothly migrate away from rubocop 0.83.0 ... which we are stuck on for
a long while now.

The script acts only when 'rubocop' gets called.